+++ code/Encoder.h
@@ -6,13 +6,13 @@
 class Encoder{
 public:
     int32_t pulse = 0;//脉冲个数，有符号
-    Encoder(uint16_t* CNT_ADD);//绑定QEI库定义的虚拟CNT
+    Encoder(long* CNT_ADD);//绑定QEI库定义的虚拟CNT
 
     void init();
     void Handler(); //更新角度，放在主程序任务轮询中断里
     void overflow_update(); //放在定时器update中断里
 private:
-    uint16_t* CNT;//编码器对应GPIO通道
+    long* CNT;//编码器对应GPIO通道
 
     int carry_flag = 0;
     uint32_t prev_cnt = 0;
